Fixing comments from PR#1497
