+++ ocs_ci/ocs/resources/storage_cluster.py
@@ -8,6 +8,8 @@ import logging
 from ocs_ci.ocs import constants, defaults
 
 from ocs_ci.ocs.resources.csv import CSV
+from ocs_ci.ocs.resources.pod import get_osd_pods
+from ocs_ci.ocs.exceptions import UnexpectedBehaviour
 from ocs_ci.ocs.resources.packagemanifest import get_selector_for_ocs_operator, PackageManifest
 from ocs_ci.utility import utils
 
@@ -327,6 +329,27 @@ def ocs_install_verification(
             ], f"{crush_rule['rule_name']} is not with type as zone"
         log.info("Verified - pool crush rule is with type: zone")
 
+
+def check_max_osds_per_node(max_number_osds=2):
+    """
+    Function to check there is less than or equal to `max_number_osds` per node
+
+    Args:
+        max_number_osds (int): upper bound
+
+    Raises:
+        UnexpectedBehaviour: if number of osds is exceeded
+
+    """
+    osd_on_node_count = {}
+    for pod in get_osd_pods():
+        node_name = pod.pod_data['spec']['nodeName']
+        osd_on_node_count[node_name] = osd_on_node_count.get(node_name, 0) + 1
+    for node_name, count in osd_on_node_count.items():
+        if count > max_number_osds:
+            raise UnexpectedBehaviour(
+                f"Node {node_name} runs {count} OSDs, maximum {max_number_osds} is allowed "
+            )
 
 def add_capacity(osd_size_capacity_requested):
     """
